Move the Editor
