+++ ui/src/components/editor/Editor/Editor.module.css
@@ -1,4 +1,4 @@
-.buildStep {
+.editor {
   width: 100%;
   height: 100%;
   display: flex;
